notification_service/pyproject.toml
[tool.poetry]
name = "notification-service"
version = "0.1.0"
description = "Сервис для отправки уведомлений через Telegram и обработки кодов подтверждения"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiogram = "^3.10.0"
fastapi = "^0.115.0"
faststream = {extras = ["rabbit"], version = "^0.5.34"}
uvicorn = "0.23.0"
python-dotenv = "1.0.0"
pydantic-settings = "^2.2.1"
typing-extensions = ">=4.8.0"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
isort = "^5.12.0"
flake8 = "^6.0.0"
mypy = "^1.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 120
lines_after_imports = 2
